Update Program.cs
With this code, you will be able to access and memorize not just one scripture but the entire Book of Mormon. I downloaded the book in PDF format initially so that anyone can use it and study through my program. I did this attempting to showcase my creativity and exceed the requirements.